I've cahnged everything over to the deterministic policy iteration and I've tried fixing the likelihood calculations and boltzmann to use the Q* values properly.
At the minute though it's not working, it doesn't seem like the boltzmann is working properly, the trajectories it's generating are of really low reward, so the true policy seems wrong and then the results are weird, the policy gives the same action for all time slots which could be something wrong with the learning algorithm or could just be because the boltzmann policy is wrong to begin with.

I'm going to push this now though so I can work from my computer.