- Implemented a gaussian mcmc step - Ran tests with gaussian mcmc step to find optimal variance (using Gelman et al. method didn't really work although I wasn't doing it properly) - Cleaned some things up - Made a few small changes based on ondrej's feedback.
With regards to vectorisation: I need to take some time to do that because when I tried to do it yesterday it made everything less efficient.

With regards to repeatedly passing around values and q_values, I do need the policy iteration to have constant access to the most up to date version of both values and q values to save time, but I could probably redesign it within a class where the values are stored as a local variable then we wouldn't need to keep passing so many parameters. I'll do this as a separate change though.