This is just the spotify_test and the npy files. Still needs a bit of clean up in places, but I'm pushing this now because it's pretty much there and I don't want to lose it.
spotify_test is responsible for handling the spotify data, displaying it over time and putting it into observation data which can be handled by my IRL algorithm. No changes are required by the IRL side of things. The only difference is that I've saved the observations and observation_times and I load these in to run BIRL on.

For now I'm not pushing any of the change to the IRL side of things because the changes are minor and I'm currently working on getting covariance matrices to work in my proposals but that would probably be too complex for this for now.

For the results that I sent in the email, I used the gaussian mcmc step with a scalar variance. I used the time dependent prior with a sigma of 0.1. Step size was 0.0035, 100,000 iterations, discount_rate of 0.1, R_max of 5 and an alpha of 2. It took 103m 25s to train the full version.

It took around half of that to train the smaller version (dec-july) that I showed Ondrej in the email.
Had to remove the learned rewards because they were 1.5 gigabytes